#2 related. Organize components
Placed loose components in "components" directory.
