+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Switch, Route, BrowserRouter as Router } from 'react-router-dom';
 
-import LandingPage from './landingPage';
-import Participation from './participation';
+import LandingPage from './components/landingPage';
+import Participation from './components/participation';
 import './App.css';
